Move runtime source primitives to parser source package
