+++ Utils.Parser.Source/IsExternalInit.cs
@@ -1,0 +1,8 @@
+namespace System.Runtime.CompilerServices;
+
+/// <summary>
+/// Provides record initialization support when targeting frameworks that do not define the type.
+/// </summary>
+internal static class IsExternalInit
+{
+}
