+++ Utils.Parser/Bootstrap/Antlr4GrammarConverter.cs
@@ -6,6 +6,7 @@ using Utils.Parser.Model;
 using Utils.Parser.Resolution;
 using Utils.Parser.Runtime;
 using System.IO;
+using Utils.Parser.Source;
 
 namespace Utils.Parser.Bootstrap;
