+++ Utils.Parser/Runtime/AlternativeScheduler.cs
@@ -1,5 +1,6 @@
 using Utils.Parser.Diagnostics;
 using Utils.Parser.Model;
+using Utils.Parser.Source;
 
 namespace Utils.Parser.Runtime;
