+++ Utils.Parser/Runtime/KeywordLookup.cs
@@ -1,4 +1,5 @@
 using Utils.Parser.Model;
+using Utils.Parser.Source;
 
 namespace Utils.Parser.Runtime;
