+++ Utils.Parser/Runtime/ParserEngine.cs
@@ -1,5 +1,6 @@
 using Utils.Parser.Model;
 using Utils.Parser.Diagnostics;
+using Utils.Parser.Source;
 
 namespace Utils.Parser.Runtime;
