+++ UtilsTest/UtilsTest.Unit.csproj
@@ -73,6 +73,7 @@
 		<ProjectReference Include="..\Utils.Parser.Diagnostics\Utils.Parser.Diagnostics.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
 		<ProjectReference Include="..\Utils.OData.Generators\Utils.OData.Generators.csproj" />
 		<ProjectReference Include="..\Utils.OData\Utils.OData.csproj" />
+		<ProjectReference Include="..\Utils.Parser.Source\Utils.Parser.Source.csproj" />
 		<ProjectReference Include="..\Utils.Parser\Utils.Parser.csproj" />
 		<ProjectReference Include="..\Utils.Expressions.CSyntax\Utils.Expressions.CSyntax.csproj" />
 		<ProjectReference Include="..\Utils.Expressions.VBSyntax\Utils.Expressions.VBSyntax.csproj" />
